added better journaling capabilities, minor internal updates & moved nop client into client.go file, too
